+++ source/Gui/UploadSimulationDialog.cpp
@@ -46,26 +46,21 @@ _UploadSimulationDialog::_UploadSimulationDialog(
     , _viewport(viewport)
     , _genomeEditorWindow(genomeEditorWindow)
 {
+    auto& settings = GlobalSettings::getInstance();
+    _share = settings.getBoolState("dialogs.upload.share", _share);
 }
 
 _UploadSimulationDialog::~_UploadSimulationDialog()
 {
+    auto& settings = GlobalSettings::getInstance();
+    settings.setBoolState("dialogs.upload.share", _share);
 }
 
 void _UploadSimulationDialog::open(NetworkResourceType resourceType, std::string const& folder)
 {
     if (NetworkService::getLoggedInUserName()) {
-        auto workspaceType = _browserWindow->getCurrentWorkspaceType();
-        if (workspaceType == WorkspaceType_AlienProject && *NetworkService::getLoggedInUserName() != "alien-project") {
-            MessageDialog::getInstance().information(
-                "Upload " + BrowserDataTypeToLowerString.at(resourceType),
-                "You are not allowed to upload to alien-project's workspace.\nPlease choose the public or private workspace in the browser.");
-            return;
-        }
-
         changeTitle("Upload " + BrowserDataTypeToLowerString.at(resourceType));
         _resourceType = resourceType;
-        _workspaceType = workspaceType;
         _folder = folder;
         _resourceName = _resourceNameByFolder[_folder];
         _resourceDescription = _resourceDescriptionByFolder[_folder];
@@ -111,9 +106,17 @@ void _UploadSimulationDialog::processIntern()
         AlienImGui::InputTextMultilineParameters()
             .hint("Description (optional)")
             .textWidth(0)
-            .height(ImGui::GetContentRegionAvail().y - StyleRepository::getInstance().scale(50.0f)),
+            .height(ImGui::GetContentRegionAvail().y - StyleRepository::getInstance().scale(70.0f)),
         _resourceDescription);
     ImGui::PopID();
+
+    AlienImGui::ToggleButton(
+        AlienImGui::ToggleButtonParameters()
+            .name("Share with public")
+            .tooltip(
+                "If true, the " + resourceTypeString + " will be visible to all users. If false, the " + resourceTypeString
+                + " will only be visible in the private workspace. This property can also be changed later if desired."),
+        _share);
 
     AlienImGui::Separator();
 
@@ -185,8 +188,9 @@ void _UploadSimulationDialog::onUpload()
         }
 
         std::string resourceId;
+        auto workspaceType = _share ? WorkspaceType_Public : WorkspaceType_Private;
         if (!NetworkService::uploadResource(
-                resourceId, _folder + _resourceName, _resourceDescription, size, numObjects, mainData, settings, statistics, _resourceType, _workspaceType)) {
+                resourceId, _folder + _resourceName, _resourceDescription, size, numObjects, mainData, settings, statistics, _resourceType, workspaceType)) {
             showMessage(
                 "Error",
                 "Failed to upload " + BrowserDataTypeToLowerString.at(_resourceType)